requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
mysql-connector-python>=8.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
